8 lines with 5 cycles are back operating

+++ include/Controller/States/StateBase.hpp
@@ -64,9 +64,9 @@ protected:
     /// Returns pointer to the extended view bny given ID, see ModuleConfig.hpp for views ID.
     ViewExtendedIf *getExtendedView(const uint8_t viewId) const;
     /// @brief Map with pointers to the standard views.
-    static std::map<uint8_t, ViewIf *> m_Views;
+    static std::map<uint32_t, ViewIf *> m_Views;
     /// @brief Map with pointers to the extended views.
-    static std::map<uint8_t, ViewExtendedIf *> m_ExtendedViews;
+    static std::map<uint32_t, ViewExtendedIf *> m_ExtendedViews;
     /// @brief Pointer to the model of the RTC. Interface to do changes.
     static ModuleModelIf *m_pModel;
     /// @brief Pointer to the model of the RTC. Interface to get state only.
